signature.m
We create a function name signature which is use to re-size the query image, convert from RGB to Gray and HSV and applying two level wavelet decomposition to find the 16 wavelet coefficient. Find the mean value of this 16 coefficient which give 16 signature of image. Generate the histogram of hue, saturation and value.Generate 18 signatures of hue, 3 for saturation and 3 for value.Load each image from database. Apply same procedure for all database images and pass results to image_retrieval.m.